Moved remaining utils.go funcs into main.go and serviceConfigWorker.go
