+++ cmd/token-push/main.go
@@ -909,3 +909,41 @@ func openDatabaseAndLoadServices(ctx context.Context) (*db.ManagedTokensDatabase
 	tracing.LogSuccessWithTrace(span, exeLogger, "Successfully opened database and loaded services")
 	return database, nil
 }
+
+// addServiceToServicesSlice checks to see if, for an experiment and its entry in the configuration, a normal service.Service can be added
+// to the services slice, or if an ExperimentOverriddenService should be added.  It then adds the resultant type that implements
+// service.Service to the services slice
+func addServiceToServicesSlice(services []service.Service, configExperiment, realExperiment, role string) []service.Service {
+	var serv service.Service
+	serviceName := realExperiment + "_" + role
+	if configExperiment != realExperiment {
+		serv = newExperimentOverriddenService(serviceName, configExperiment)
+	} else {
+		serv = service.NewService(serviceName)
+	}
+	services = append(services, serv)
+	return services
+}
+
+// getDevEnvironment first checks the environment variable MANAGED_TOKENS_DEV_ENVIRONMENT for the devEnvironment, then the configuration file.
+// If it finds neither are set, it returns the default global setting.  This logic is handled by the underlying logic in the
+// viper library
+func getDevEnvironmentLabel() string {
+	// For devs, this variable can be set to differentiate between dev and prod for metrics, for example
+	viper.SetDefault("devEnvironmentLabel", devEnvironmentLabelDefault)
+	viper.BindEnv("devEnvironmentLabel", "MANAGED_TOKENS_DEV_ENVIRONMENT_LABEL")
+	return viper.GetString("devEnvironmentLabel")
+}
+
+// getPrometheusJobName gets the job name by parsing the configuration and the devEnvironment
+func getPrometheusJobName() string {
+	defaultJobName := "managed_tokens"
+	jobName := viper.GetString("prometheus.jobname")
+	if jobName == "" {
+		jobName = defaultJobName
+	}
+	if devEnvironmentLabel == devEnvironmentLabelDefault {
+		return jobName
+	}
+	return fmt.Sprintf("%s_%s", jobName, devEnvironmentLabel)
+}
