+++ cmd/token-push/serviceConfigWorker.go
@@ -1,0 +1,99 @@
+// COPYRIGHT 2024 FERMI NATIONAL ACCELERATOR LABORATORY
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+//
+// You may obtain a copy of the License at
+//
+//	http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package main
+
+import (
+	"context"
+	"reflect"
+	"runtime"
+
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/attribute"
+
+	"github.com/fermitools/managed-tokens/internal/utils"
+	"github.com/fermitools/managed-tokens/internal/worker"
+)
+
+// chansForWorkers is an interface that defines methods for obtaining channels
+// used by worker.Workers. It includes methods to get channels for service
+// configurations and success reporting.
+type chansForWorkers interface {
+	GetServiceConfigChan() chan<- *worker.Config
+	GetSuccessChan() <-chan worker.SuccessReporter
+}
+
+// startServiceConfigWorkerForProcessing starts up a worker using the provided workerFunc, gives it a set of channels to receive *worker.Configs
+// and send notification.Notifications on, and sends *worker.Configs to the worker
+func startServiceConfigWorkerForProcessing(ctx context.Context, workerFunc worker.Worker,
+	serviceConfigs map[string]*worker.Config, timeoutCheckKey timeoutKey, disableNotifications bool) chansForWorkers {
+	// Channels, context, and worker for getting kerberos tickets
+	ctx, span := otel.GetTracerProvider().Tracer("token-push").Start(ctx, "startServiceConfigWorkerForProcessing")
+	span.SetAttributes(
+		attribute.KeyValue{
+			Key:   "workerFunc",
+			Value: attribute.StringValue(runtime.FuncForPC(reflect.ValueOf(workerFunc).Pointer()).Name()),
+		},
+	)
+	defer span.End()
+
+	var useCtx context.Context
+
+	channels := worker.NewChannelsForWorkers(len(serviceConfigs))
+
+	if !disableNotifications {
+		startListenerOnWorkerNotificationChans(ctx, channels.GetNotificationsChan())
+	}
+
+	if timeout, ok := timeouts[timeoutCheckKey]; ok {
+		useCtx = utils.ContextWithOverrideTimeout(ctx, timeout)
+	} else {
+		useCtx = ctx
+	}
+	go workerFunc(useCtx, channels)
+	if len(serviceConfigs) > 0 { // We add this check because if there are no serviceConfigs, don't load them into any channel
+		serviceConfigSlice := make([]*worker.Config, 0, len(serviceConfigs))
+		for _, sc := range serviceConfigs {
+			serviceConfigSlice = append(serviceConfigSlice, sc)
+		}
+		loadServiceConfigsIntoChannel(channels.GetServiceConfigChan(), serviceConfigSlice)
+	}
+	return channels
+}
+
+// loadServiceConfigsIntoChannel loads *worker.Config objects into a channel, usually for use by a worker, and then closes the channel
+func loadServiceConfigsIntoChannel(chanToLoad chan<- *worker.Config, serviceConfigSlice []*worker.Config) {
+	defer close(chanToLoad)
+	for _, sc := range serviceConfigSlice {
+		chanToLoad <- sc
+	}
+}
+
+// removeFailedServiceConfigs reads the worker.SuccessReporter chan from the passed in worker.ChannelsForWorkers object, and
+// removes any *worker.Config objects from the passed in serviceConfigs map.  It returns a slice of the *worker.Configs that
+// were removed
+func removeFailedServiceConfigs(chans chansForWorkers, serviceConfigs map[string]*worker.Config) []*worker.Config {
+	failedConfigs := make([]*worker.Config, 0, len(serviceConfigs))
+	for workerSuccess := range chans.GetSuccessChan() {
+		if !workerSuccess.GetSuccess() {
+			exeLogger.WithField(
+				"service", getServiceName(workerSuccess.GetService()),
+			).Debug("Removing serviceConfig from list of configs to use")
+			failedConfigs = append(failedConfigs, serviceConfigs[getServiceName(workerSuccess.GetService())])
+			delete(serviceConfigs, getServiceName(workerSuccess.GetService()))
+		}
+	}
+	return failedConfigs
+}
